move create from profile template action to export group (#165939)
