+++ src/vs/workbench/contrib/userDataProfile/browser/userDataProfileActions.ts
@@ -3,17 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { CancellationToken } from 'vs/base/common/cancellation';
-import { DisposableStore } from 'vs/base/common/lifecycle';
 import { localize } from 'vs/nls';
 import { Action2, IMenuService, registerAction2 } from 'vs/platform/actions/common/actions';
-import { IDialogService, IFileDialogService } from 'vs/platform/dialogs/common/dialogs';
-import { IFileService } from 'vs/platform/files/common/files';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { INotificationService } from 'vs/platform/notification/common/notification';
 import { QuickPickItem, IQuickInputService, IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
-import { asJson, asText, IRequestService } from 'vs/platform/request/common/request';
-import { IUserDataProfileTemplate, isUserDataProfileTemplate, IUserDataProfileManagementService, IUserDataProfileImportExportService, PROFILES_CATEGORY, ManageProfilesSubMenu, IUserDataProfileService, PROFILES_ENABLEMENT_CONTEXT, HAS_PROFILES_CONTEXT, MANAGE_PROFILES_ACTION_ID, PROFILE_FILTER } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
+import { IUserDataProfileManagementService, PROFILES_CATEGORY, ManageProfilesSubMenu, IUserDataProfileService, PROFILES_ENABLEMENT_CONTEXT, HAS_PROFILES_CONTEXT, MANAGE_PROFILES_ACTION_ID } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
 import { IUserDataProfile, IUserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { Categories } from 'vs/platform/action/common/actionCommonCategories';
 import { ContextKeyExpr, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
@@ -141,9 +136,6 @@ registerAction2(class CreateProfileAction extends Action2 {
 			}, {
 				id: CreateFromCurrentProfileAction.ID,
 				label: CreateFromCurrentProfileAction.TITLE.value,
-			}, {
-				id: CreateFromTemplateAction.ID,
-				label: CreateFromTemplateAction.TITLE.value,
 			}, {
 				type: 'separator',
 			}, {
@@ -385,130 +377,6 @@ registerAction2(class SwitchProfileAction extends Action2 {
 		}
 	}
 });
-
-class ImportProfileAction extends Action2 {
-	static readonly ID = 'workbench.profiles.actions.importProfile';
-	static readonly TITLE = {
-		value: localize('import profile', "Import from ..."),
-		original: 'Import...'
-	};
-	constructor() {
-		super({
-			id: ImportProfileAction.ID,
-			title: ImportProfileAction.TITLE,
-			category: PROFILES_CATEGORY,
-			f1: true,
-			precondition: PROFILES_ENABLEMENT_CONTEXT?.negate(),
-		});
-	}
-
-	async run(accessor: ServicesAccessor) {
-		const fileDialogService = accessor.get(IFileDialogService);
-		const quickInputService = accessor.get(IQuickInputService);
-		const fileService = accessor.get(IFileService);
-		const requestService = accessor.get(IRequestService);
-		const userDataProfileImportExportService = accessor.get(IUserDataProfileImportExportService);
-		const dialogService = accessor.get(IDialogService);
-		const contextKeyService = accessor.get(IContextKeyService);
-		const notificationService = accessor.get(INotificationService);
-
-		const isSettingProfilesEnabled = contextKeyService.contextMatchesRules(PROFILES_ENABLEMENT_CONTEXT);
-
-		if (!isSettingProfilesEnabled) {
-			if (!(await dialogService.confirm({
-				title: localize('import profile title', "Import Settings from a Profile"),
-				message: localize('confiirmation message', "This will replace your current settings. Are you sure you want to continue?"),
-			})).confirmed) {
-				return;
-			}
-		}
-
-		const disposables = new DisposableStore();
-		const quickPick = disposables.add(quickInputService.createQuickPick());
-		const updateQuickPickItems = (value?: string) => {
-			const selectFromFileItem: IQuickPickItem = { label: isSettingProfilesEnabled ? localize('select from file', "Select Settings Profile template file") : localize('import from file', "Import from profile file") };
-			quickPick.items = value ? [{ label: isSettingProfilesEnabled ? localize('select from url', "Create from template URL") : localize('import from url', "Import from URL"), description: quickPick.value }, selectFromFileItem] : [selectFromFileItem];
-		};
-		quickPick.title = isSettingProfilesEnabled ? localize('create from profile template quick pick title', "Create from Settings Profile Template") : localize('import profile quick pick title', "Import Settings from a Profile");
-		quickPick.placeholder = isSettingProfilesEnabled ? localize('create from profile template placeholder', "Provide a template URL or Select a template file") : localize('import profile placeholder', "Provide profile URL or select profile file to import");
-		quickPick.ignoreFocusOut = true;
-		disposables.add(quickPick.onDidChangeValue(updateQuickPickItems));
-		updateQuickPickItems();
-		quickPick.matchOnLabel = false;
-		quickPick.matchOnDescription = false;
-		disposables.add(quickPick.onDidAccept(async () => {
-			try {
-				quickPick.hide();
-				const profile = quickPick.selectedItems[0].description ? await this.getProfileFromURL(quickPick.value, requestService) : await this.getProfileFromFileSystem(fileDialogService, fileService);
-				if (profile) {
-					if (isSettingProfilesEnabled) {
-						await userDataProfileImportExportService.importProfile(profile);
-					} else {
-						await userDataProfileImportExportService.setProfile(profile);
-					}
-				}
-			} catch (error) {
-				notificationService.error(error);
-			}
-		}));
-		disposables.add(quickPick.onDidHide(() => disposables.dispose()));
-		quickPick.show();
-	}
-
-	private async getProfileFromFileSystem(fileDialogService: IFileDialogService, fileService: IFileService): Promise<IUserDataProfileTemplate | null> {
-		const profileLocation = await fileDialogService.showOpenDialog({
-			canSelectFolders: false,
-			canSelectFiles: true,
-			canSelectMany: false,
-			filters: PROFILE_FILTER,
-			title: localize('import profile dialog', "Import Profile"),
-		});
-		if (!profileLocation) {
-			return null;
-		}
-		const content = (await fileService.readFile(profileLocation[0])).value.toString();
-		const parsed = JSON.parse(content);
-		return isUserDataProfileTemplate(parsed) ? parsed : null;
-	}
-
-	private async getProfileFromURL(url: string, requestService: IRequestService): Promise<IUserDataProfileTemplate | null> {
-		const options = { type: 'GET', url };
-		const context = await requestService.request(options, CancellationToken.None);
-		if (context.res.statusCode === 200) {
-			const result = await asJson(context);
-			return isUserDataProfileTemplate(result) ? result : null;
-		} else {
-			const message = await asText(context);
-			throw new Error(`Expected 200, got back ${context.res.statusCode} instead.\n\n${message}`);
-		}
-	}
-
-}
-registerAction2(ImportProfileAction);
-
-class CreateFromTemplateAction extends Action2 {
-	static readonly ID = 'workbench.profiles.actions.createFromTemplate';
-	static readonly TITLE = {
-		value: localize('create from template profile', "Create from a Settings Profile Template..."),
-		original: 'Create from a Settings Profile Template...'
-	};
-	constructor() {
-		super({
-			id: CreateFromTemplateAction.ID,
-			title: CreateFromTemplateAction.TITLE,
-			category: PROFILES_CATEGORY,
-			f1: true,
-			precondition: PROFILES_ENABLEMENT_CONTEXT,
-		});
-	}
-
-	async run(accessor: ServicesAccessor) {
-		return accessor.get(ICommandService).executeCommand(ImportProfileAction.ID);
-	}
-
-}
-
-registerAction2(CreateFromTemplateAction);
 
 // Developer Actions
